[level 0] Title: 두 수의 나눗셈, Time: 0.01 ms, Memory: 4.2 MB -BaekjoonHub

+++ 프로그래머스/0/120806. 두 수의 나눗셈/두 수의 나눗셈.c
@@ -6,7 +6,7 @@ int solution(int num1, int num2) {
     int answer = 0;
    
     
-    answer  = (double)num1/num2 * 1000;
+    answer  = (float)num1/num2 * 1000;
         
     return answer;
 }
